requirements.txt
Django~=4.0.4
psycopg2-binary
djangorestframework